requirements.txt
Babel==2.5.3
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Cors==3.0.4
Flask-Dance==0.14.0
Flask-JWT-Extended==3.9.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
MarkupSafe==1.0
marshmallow==2.15.2
marshmallow-sqlalchemy==0.13.2
oauthlib==2.0.7
passlib==1.7.1
psycopg2==2.7.4
PyJWT==1.6.1
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.7
SQLAlchemy-Utils==0.33.3
urllib3==1.22
URLObject==2.4.3
Werkzeug==0.14.1
WTForms==2.1
